+++ packages/drawing/src/lib/directives/drawing-manager-trigger.ts
@@ -1,4 +1,4 @@
-import { AgmMap } from '@wkaskie/agm-core';
+import { AgmMap } from '@wafka/agm-core';
 import { AfterViewInit, Directive, Host, Input, OnDestroy } from '@angular/core';
 import { first } from 'rxjs/operators';
 import { AgmDrawingManager } from './drawing-manager';