+++ packages/markerclusterer/src/lib/directives/marker-cluster.ts
@@ -2,7 +2,7 @@ import { Directive, EventEmitter, Input, OnChanges, OnDestroy, OnInit, Output, S
 
 import { Subscription } from 'rxjs';
 
-import { InfoWindowManager, MarkerManager } from '@wkaskie/agm-core';
+import { InfoWindowManager, MarkerManager } from '@wafka/agm-core';
 import { ClusterIconStyle, MarkerClustererOptions } from '@google/markerclustererplus';
 import { Calculator } from '@google/markerclustererplus/dist/markerclusterer';
 import { ClusterManager } from '../services/managers/cluster-manager';