+++ packages/markerclusterer/src/lib/marker-clusterer.module.ts
@@ -1,4 +1,4 @@
-import { AgmCoreModule } from '@wkaskie/agm-core';
+import { AgmCoreModule } from '@wafka/agm-core';
 import { NgModule } from '@angular/core';
 import { AgmMarkerCluster } from './directives/marker-cluster';