+++ packages/markerclusterer/src/lib/services/managers/cluster-manager.ts
@@ -1,4 +1,4 @@
-import { AgmMarker, GoogleMapsAPIWrapper, MarkerManager } from '@wkaskie/agm-core';
+import { AgmMarker, GoogleMapsAPIWrapper, MarkerManager } from '@wafka/agm-core';
 import { Injectable, NgZone } from '@angular/core';
 import { MarkerClustererOptions } from '@google/markerclustererplus';
 import MarkerClusterer from '@google/markerclustererplus';